requirements.txt
pyyaml
python-dotenv
crewai
crewai_tools